requirements.txt
asgiref==3.4.1
Django==3.2.7
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-filter==21.1
djangorestframework==3.12.4
flake8==3.9.2
mccabe==0.6.1
Pillow==8.3.2
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
sqlparse==0.4.2
